Make performer page tabbed
